Moved fcns for persisting showNullValues to a seperate file
